files moved and added Trein class
