+++ RadvanfortreinAndroid/src/app/app.module.ts
@@ -5,9 +5,9 @@ import { FormsModule, ReactiveFormsModule} from '@angular/forms';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './/app-routing.module';
-import { StartComponent } from './domain/start/start.component';
-import { InzetComponent } from './domain/inzet/inzet.component';
-import { ResultaatComponent } from './domain/resultaat/resultaat.component';
+import { StartComponent } from './start/start.component';
+import { InzetComponent } from './inzet/inzet.component';
+import { ResultaatComponent } from './resultaat/resultaat.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { LayoutModule } from '@angular/cdk/layout';
 import { MatToolbarModule, MatInputModule, MatButtonModule, 
@@ -15,7 +15,7 @@ import { MatToolbarModule, MatInputModule, MatButtonModule,
   MatTabsModule, MatStepperModule, MatFormFieldModule, MatSliderModule, MatRadioModule, MatTableModule
  } from '@angular/material';
 import { TabsComponent } from './material/tabs/tabs.component';
-import { TabelComponent } from './domain/inzet/tabel/tabel.component';
+import { TabelComponent } from './inzet/tabel/tabel.component';
 
 @NgModule({
   declarations: [
