+++ RadvanfortreinAndroid/src/app/domain/Trein/trein.spec.ts
@@ -1,0 +1,7 @@
+import { Trein } from './trein';
+
+describe('Trein', () => {
+  it('should create an instance', () => {
+    expect(new Trein()).toBeTruthy();
+  });
+});
