+++ RadvanfortreinAndroid/src/app/resultaat/resultaat.component.ts
