+++ RadvanfortreinAndroid/src/app/start/start.component.ts
